Move logic to a form-behavior that can be installed on native forms
