+++ components/form/demo/form-behavior.html
@@ -1,0 +1,82 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+	<meta name="viewport" content="width=device-width, initial-scale=1.0">
+	<meta http-equiv="X-UA-Compatible" content="ie=edge">
+	<meta charset="UTF-8">
+	<link rel="stylesheet" href="../../demo/styles.css" type="text/css">
+	<script src="/node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
+	<script type="module">
+		import '../../demo/demo-page.js';
+		import '../../button/button.js';
+		import '../../inputs/input-text.js';
+	</script>
+</head>
+
+<body unresolved>
+
+	<d2l-demo-page page-title="d2l-form">
+
+		<h2>Default</h2>
+		<d2l-demo-snippet>
+			<template>
+				<form method="POST">
+					<h1>My Form</h1>
+					<h2>Primary</h2>
+					<fieldset>
+						<legend>Choose your favorite monster</legend>
+
+						<input type="radio" id="kraken" name="monster" value="oh no">
+						<label for="kraken">Kraken</label><br />
+
+						<input type="radio" id="sasquatch" name="monster" value="wow">
+						<label for="sasquatch">Sasquatch</label><br />
+
+					</fieldset>
+					<input type="checkbox" id="kraken" name="checkers" value="red-black">
+					<div>
+						<label for="name">Name (4 to 8 characters):</label>
+						<input type="text" id="name" name="name" required minlength="4" maxlength="8" size="10">
+
+						<div>
+							<select name="pets" id="pet-select">
+								<option value="">--Please choose an option--</option>
+								<option value="dog">Dog</option>
+								<option value="cat">Cat</option>
+								<option value="hamster">Hamster</option>
+								<option value="parrot">Parrot</option>
+								<option value="spider">Spider</option>
+								<option value="goldfish">Goldfish</option>
+							</select>
+						</div>
+					</div>
+					<d2l-input-text name="custom-input" required></d2l-input-text>
+					<div>
+						<h2>Secondary</h2>
+						<label for="story">Tell us your story:</label>
+						<textarea id="story" name="story" rows="5" cols="33">It was a dark and stormy night...</textarea>
+						<div>
+							<input type="range" id="b" name="b" value="50" /> +
+							<input type="number" id="a" name="a" value="10" /> =
+							<output name="result" for="a b">60</output>
+						</div>
+					</div>
+					<d2l-button type="submit" name="action">Update</d2l-button>
+					<d2l-button type="submit" name="action" value="delete">Delete</d2l-button>
+					<button type="submit" name="abc">Wow</button>
+				</form>
+				<script type="module">
+					import {FormBehavior } from  '../form-behavior.js';
+					const form = document.querySelector('form');
+					const behavior = new FormBehavior();
+					behavior.install(form);
+				</script>
+			</template>
+		</d2l-demo-snippet>
+
+	</d2l-demo-page>
+
+</body>
+
+</html>
